SSH users are now created with public keys, tested with cukes
It appears that running `chef-solo` for a single .json will delete
all cookbooks in `/cookbooks` that are NOT related to the Runlist of the
given .json. It seems there are only two ways to avoid this churn:

1. Run chef-solo in --legacy mode

But who knows how long --legacy will stick around for?

2. Combine all of the .jsons together into one

But then it's harder to just update a single Chef recipe/deployment

I wonder if there's a way to automatically smoosh all of the .jsons
into a single .json... #interesting